refactor: extract LatexSuiteSettings into own file
